add importmap to the config
